backend/requirements.txt
fastapi
uvicorn[standard]
asyncpg
python-dotenv
passlib[bcrypt]
python-jose[cryptography]